requirements.txt
shapely
pygame
pyyaml